+++ Sources/VCore/Views/SwiftUI/Plain List/PlainListUIModel.swift
@@ -1,0 +1,20 @@
+//
+//  PlainListUIModel.swift
+//  VCore
+//
+//  Created by Vakhtang Kontridze on 27.11.23.
+//
+
+import SwiftUI
+
+// MARK: - Plain List UI Model
+/// Model that describes UI.
+public struct PlainListUIModel {
+    // MARK: Properties
+    /// Row background color.
+    public var rowBackgroundColor: Color = .clear
+
+    // MARK: Initializers
+    /// Initializes UI model with default values.
+    public init() {}
+}